Do not ENABLE_OPTIONAL_FEATURES in consuming applications.
`config/environment.js` is used by ember-cli to allow the addon to
provide default configuration to the consuming app.  In this case, the
configuration here is for the addon's internal development so it belongs
in `tests/dummy/config/environment.js` instead.
